newsletter_service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.3
pydantic-settings==2.2.1
celery==5.3.6
redis==5.0.1
requests==2.31.0
sendgrid==6.11.0
boto3==1.34.51
jinja2==3.1.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==7.0.0
